requirements.txt
crewai
crewai-tools
langchain-community
langchain
langchain-huggingface
langchain-pinecone
pinecone-client
sentence-transformers
python-dotenv
tavily-python
streamlit
litellm
torchvision
